+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
+## [Unreleased](//github.com/opentable/sous/compare/0.2.1...HEAD)
+
+### Added
+- Conflicting GDM updates now retry, up to the number of deployments in their manifest.
+
 ## [0.2.1](//github.com/opentable/sous/compare/0.2.0...0.2.1)
 
 ### Added